+++ 2 semester/3 home work/2-3-4/2-3-4/StackElement.cs
@@ -8,11 +8,11 @@ namespace _2_3_4
 {
     public class StackElement
     {
-        public int Element { get; set; }
+        public float Element { get; set; }
 
         public StackElement Next { get; set; }
 
-        public StackElement(StackElement next, int value)
+        public StackElement(StackElement next, float value)
         {
             this.Next = next;
             this.Element = value;
